requirements.txt
psycopg2==2.5.4
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
djangorestframework==2.4.4
gunicorn==19.1.1
static3==0.5.1
wsgiref==0.1.2
